# Perf: KZG verification in a single point